requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bitarray==3.7.1
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.0.1
defusedxml==0.7.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.104.1
fonttools==4.60.0
frozenlist==1.7.0
graphviz==0.21
h11==0.16.0
hexbytes==1.3.1
idna==3.10
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
lru-dict==1.3.0
matplotlib==3.10.6
multidict==6.6.4
narwhals==2.5.0
numpy==1.26.1
odfpy==1.4.1
packaging==25.0
pandas==2.1.3
parsimonious==0.10.0
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
protobuf==6.32.1
pycryptodome==3.23.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyunormalize==16.0.0
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.31.0
rlp==4.1.0
rpds-py==0.27.1
scikit-learn==1.3.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tabulate==0.9.0
threadpoolctl==3.6.0
toolz==1.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0.post1
web3==6.10.0
websockets==15.0.1
xgboost==2.0.3
yarl==1.20.1
